Appending data to the Google sheet.
Includes adding a Sheet ID env var to the function in Terraform and
accessing the Sheet via the API. Uses the Default Google Application
Authentication. Basically, API library dectects that it is running in
the context of a Cloud Function and uses the default Service Account to
autenticate. That Service Account was added as an editor to the Sheet.
Super simple.